`refactor`: revert `toLong` to `toInt` due to require field from gql is `Int`

+++ features/content/createpost/src/main/java/com/tokopedia/createpost/view/presenter/CreatePostPresenter.kt
@@ -110,7 +110,7 @@ class CreatePostPresenter @Inject constructor(
     override fun getFollowersCount(isAffiliateType: Boolean) {
         if (isAffiliateType) {
             getProfileHeaderUseCase.execute(
-                    GetProfileHeaderUseCase.createRequestParams(userSession.userId.toLong()),
+                    GetProfileHeaderUseCase.createRequestParams(userSession.userId.toInt()),
                     object : Subscriber<GraphqlResponse>() {
                         override fun onNext(t: GraphqlResponse?) {
                             followersCount = t?.let(::getFollowersCount)
